@@ -2,11 +2,10 @@ import axios from 'axios';
 import * as types from './actionTypes';
 import { searchSuccess } from './userAction';
 /**
- *
- *
- * @export
+ * 
+ * @export 
  * @param {any} documents
- * @returns
+ * @returns {object} dispatch
  */
 export function listDocs(documents) {
   return {
@@ -19,7 +18,7 @@ export function listDocs(documents) {
  *
  * @export
  * @param {any} id
- * @returns
+ * @returns {object} dispatch
  */
 export function deleteDocs(id) {
   return {
@@ -31,8 +30,8 @@ export function deleteDocs(id) {
  *
  *
  * @export
- * @param {any} id
- * @returns
+ * @param {any} document
+ * @returns {object} dispatch
  */
 export function updateDocs(document) {
   return {
@@ -45,7 +44,7 @@ export function updateDocs(document) {
  *
  * @export
  * @param {any} document
- * @returns
+ * @returns {object} dispatch
  */
 export function fetchDoc(document) {
   return {
@@ -58,7 +57,7 @@ export function fetchDoc(document) {
  *
  * @export
  * @param {any} document
- * @returns
+ * @returns {object} dispatch
  */
 export function create(document) {
   return {
@@ -72,110 +71,113 @@ export function create(document) {
  * @export
  * @param {any} limit
  * @param {any} offset
- * @returns
+ * @returns {object} dispatch
  */
-
 export function doclist(limit, offset) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch =>
-  axios.get(`/api/v1/documents/?limit=${limit || 10}
+    axios.get(`/api/v1/documents/?limit=${limit || 10}
   &offset=${offset || 0}`)
-  .then((response) => {
-    dispatch(listDocs(response.data));
-    dispatch({ type: 'Error' });
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error.response.data });
-  });
+      .then((response) => {
+        dispatch(listDocs(response.data));
+        // dispatch({ type: 'Error' });
+      })
+      .catch((error) => {
+        dispatch({ type: 'Error', error: error.response.data });
+      });
 }
 /**
  *
  *
  * @export
  * @param {any} docId
- * @returns
+ * @returns {object} dispatch
  */
 export function deleteDoc(docId) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch => axios.delete(`api/v1/documents/${docId}`)
   .then((response) => {//eslint-disable-line
-    dispatch(deleteDocs(docId));
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error.response.data });
-  });
+      dispatch(deleteDocs(docId));
+    })
+    .catch((error) => {
+      dispatch({ type: 'Error', error: error.response.data });
+    });
 }
 /**
  *
  *
  * @export
  * @param {any} document
- * @returns
+ * @returns {object} dispatch
  */
 export function updateDoc(document) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch => axios.put(`/api/v1/documents/${document.id}`, document)
-  .then((response) => {
-    dispatch(updateDocs(response.data));
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error.response.data });
-  });
+    .then((response) => {
+      dispatch(updateDocs(response.data));
+    })
+    .catch((error) => {
+      dispatch({ type: 'Error', error: error.response.data });
+    });
 }
 /**
  *
  *
  * @export
  * @param {any} id
- * @returns
+ * @returns {object} dispatch
  */
 export function findDoc(id) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch => axios.get(`/api/v1/documents/${id}`)
-  .then((response) => {
-    dispatch(fetchDoc(response.data.documents));
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error });
-  });
+    .then((response) => {
+      dispatch(fetchDoc(response.data.documents));
+    })
+    .catch((error) => {
+      if (error.response.status === 401) {
+        dispatch({ type: types.UNAUTHORIZED, error: error.response.data });
+      } else {
+        dispatch({ type: 'Error', error: error.response.data });
+      }
+    });
 }
 /**
  *
  *
  * @export
  * @param {any} document
- * @returns
+ * @returns {object} dispatch
  */
 export function createDoc(document) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch => axios.post('/api/v1/documents', document)
-  .then((response) => {
-    dispatch(create(response.data));
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error.response.data });
-  });
+    .then((response) => {
+      dispatch(create(response.data));
+    })
+    .catch((error) => {
+      dispatch({ type: 'Error', error: error.response.data });
+    });
 }
 /**
  *
  *
  * @export
  * @param {any} document
- * @returns
+ * @returns {object} dispatch
  */
 export function searchDoc(document) {
   const token = window.localStorage.getItem('token');
   axios.defaults.headers.common.Authorization = token;
   return dispatch => axios.get(`/api/v1/search/documents/?title=${document}`)
-  .then((response) => {
-    dispatch(searchSuccess(response.data.result.rows, response.data.metadata));
-  })
-  .catch((error) => {
-    dispatch({ type: 'Error', error: error.response.data });
-  });
+    .then((response) => {
+      dispatch(searchSuccess(response.data.result.rows, response.data.metadata));
+    })
+    .catch((error) => {
+      dispatch({ type: 'Error', error: error.response.data });
+    });
 }
