@@ -1,14 +1,17 @@
 import React from 'react';
+
 const Footer = () =>
-          <footer className="page-footer">
-          <div className="container" />
-          <div className="footer-copyright">
-            <div className="container">
+  (<footer className="page-footer">
+    <div className="container" />
+    <div className="footer-copyright">
+      <div className="container">
             © DocMan 2017
-            <a className="grey-text text-lighten-4 right"
-            href="#!">More Links</a>
-            </div>
-          </div>
-        </footer>
+        <a
+          className="grey-text text-lighten-4 right"
+          href="#!"
+        >More Links</a>
+      </div>
+    </div>
+  </footer>)
     ;
 export default Footer;
