@@ -1,10 +1,16 @@
-import React, { PropTypes } from 'react';
+import React from 'react';
+import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { browserHistory, Link } from 'react-router';
 import * as userActions from '../../actions/userAction';
 
-
+/**
+ * 
+ * 
+ * @class Nav
+ * @extends {React.Component}
+ */
 class Nav extends React.Component {
   /**
    * Creates an instance of Nav.
@@ -25,6 +31,7 @@ class Nav extends React.Component {
    *
    *
    * @memberof Nav
+   * @returns {null} null
    */
   componentDidMount() {
     this.props.actions.activeUser();
@@ -32,7 +39,7 @@ class Nav extends React.Component {
   /**
    * 
    * 
-   * 
+   * @returns {null} null
    * @memberof Nav
    */
   componentWillUnmount() {
@@ -41,7 +48,7 @@ class Nav extends React.Component {
   /**
    *
    *
-   *
+   * @returns {null} null
    * @memberof Nav
    */
   handleClick() {
@@ -49,38 +56,55 @@ class Nav extends React.Component {
     this.props.actions.logout();
     browserHistory.push('/');
   }
-/**
+  /**
  *
  *
- * @returns
+ * @returns {null} null
  *
  * @memberof Nav
  */
   render() {
     const { authUser } = this.props;
     return (
-    <nav>
-    <div className="nav-wrapper" id="nav">
-      <Link to = "/home" className="brand-logo" id="app-name">Doc</Link>
-      <ul id="nav-mobile" className="right hide-on-med-and-down">
-       <li className= "orange-text text-darken-4"><b>Hi {authUser.name}! </b></li>
-    {authUser.roleId === 1 ? <li><Link to="/users">Users</Link></li> : ''}
-         <li> <Link to="/documents">Documents</Link></li>
-          <li><Link to="/documents/create">New Document</Link>
-</li>
-        <li><Link to="/profile"> Edit Profile?</Link></li>
-        <li><a className= "right" onClick= {this.handleClick} >Logout</a></li>
-      </ul>
-    </div>
-  </nav>
+      <nav>
+        <div
+          className="nav-wrapper"
+          id="nav"
+        >
+          <Link to="/home" className="brand-logo" id="app-name">Doc</Link>
+          <ul id="nav-mobile" className="right hide-on-med-and-down">
+            <li className="orange-text text-darken-4"><b>Hi {authUser.name}! </b></li>
+            {authUser.roleId === 1 ? <li><Link to="/users">Users</Link></li> : ''}
+            <li> <Link to="/documents">Documents</Link></li>
+            <li><Link to="/documents/create" id="newDocument">New Document</Link></li>
+            <li><Link to="/profile" id="profile"> Edit Profile?</Link></li>
+            <li>
+              <a
+                role="link"
+                tabIndex="0"
+                className="right"
+                id="logout"
+                onClick={this.handleClick}
+              >
+                Logout
+              </a>
+            </li>
+          </ul>
+        </div>
+      </nav>
     );
   }
 }
+Nav.propTypes = {
+  authUser: PropTypes.object.isRequired,
+  actions: PropTypes.object.isRequired,
+};
+
 /**
  *
  *
  * @param {any} state
- * @returns
+ * @returns {object} state
  */
 function mapStateToProps(state) {
   return {
@@ -91,7 +115,7 @@ function mapStateToProps(state) {
  *
  *
  * @param {any} dispatch
- * @returns
+ * @returns {object} object
  */
 function mapDispatchToProps(dispatch) {
   return {
