@@ -1,9 +1,15 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { browserHistory, Link } from 'react-router';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import * as userActions from '../../actions/userAction';
-
+/**
+ * 
+ * 
+ * @class Login
+ * @extends {React.Component}
+ */
 class Login extends React.Component {
   /**
    * Creates an instance of Login.
@@ -21,34 +27,34 @@ class Login extends React.Component {
     this.handleChange = this.handleChange.bind(this);
     this.handleSubmit = this.handleSubmit.bind(this);
   }
-/**
- * 
- * 
- * 
+  /**
+ *
+ *
+ *
  * @memberof Login
  */
   componentDidMount() {
-    $(document).ready(function () {
+    $(document).ready(() => {
       $('.slider').slider();
     });
   }
   /**
-   * 
-   * 
-   * @param {any} nextProps 
-   * 
+   *
+   *
+   * @param {any} nextProps
+   * @returns {state} state
    * @memberof Login
    */
   componentWillReceiveProps(nextProps) {
-    if(!this.props.user.isAuthenticated && nextProps.user.isAuthenticated) {
-     return browserHistory.push('/home'); 
+    if (!this.props.user.isAuthenticated && nextProps.user.isAuthenticated) {
+      return browserHistory.push('/home');
     }
-    return this.setState({error: nextProps.user.error})
+    return this.setState({ error: nextProps.user.error });
   }
   /**
    * 
    * 
-   * 
+   * @returns {null} null
    * @memberof Login
    */
   componentWillUnmount() {
@@ -56,7 +62,7 @@ class Login extends React.Component {
   }
   /**
    *
-   *
+   * @returns {null} null
    * @param {any} event
    *
    * @memberof Login
@@ -72,7 +78,7 @@ class Login extends React.Component {
    *
    *
    * @param {any} event
-   * @returns
+   * @returns {null} null
    *
    * @memberof Login
    */
@@ -86,7 +92,7 @@ class Login extends React.Component {
   /**
    *
    *
-   * @returns
+   * @returns {null} null
    *
    * @memberof Login
    */
@@ -95,46 +101,61 @@ class Login extends React.Component {
       <div className="slider fullscreen">
         <ul className="slides">
           <li>
-            <img src="/images/Landing.png" />
+            <img src="/images/Landing.png" alt="" />
             <div className="caption center-align">
               <h3 id="landing-text">Doc Manager</h3>
-              <h5 className="light grey-text text-lighten-3">Document Management System</h5>
+              <h5
+                className="light grey-text text-lighten-3"
+              >
+                Document Management System
+              </h5>
               <div className="row">
                 <h4>Login</h4>
-                <form className="col s12 offset-s3" onSubmit={this.handleSubmit}>
+                <form
+                  className="col s12 offset-s3"
+                  onSubmit={this.handleSubmit}
+                >
                   <div className="row">
                     <div className="input-field col s6 ">
-                      <input placeholder="Email"
+                      <input
+                        placeholder="Email"
+                        required="required"
                         name="email"
                         id="email"
                         type="email"
                         className="validate input"
                         value={this.state.email}
-                        onChange={this.handleChange} />
+                        onChange={this.handleChange}
+                      />
                     </div>
-                    </div>
-                    <div className = "row">
+                  </div>
+                  <div className="row">
                     <div className="input-field col s6">
-                      <input placeholder="Password"
+                      <input
+                        placeholder="Password"
+                        required="required"
                         name="password"
                         id="password"
                         type="password"
                         className="validate"
                         value={this.state.password}
-                        onChange={this.handleChange} />
+                        onChange={this.handleChange}
+                      />
                     </div>
                   </div>
-                  <div className= "row col s6">
-                  <button className="btn waves-effect waves-light"
-                    id= "button"
-                    type="submit"
-                    name="action"
-                    value="submit">
-                    Login
-<i className="material-icons right">
-                      send
-                  </i>
-                  </button>
+                  <div className="row col s6">
+                    <button
+                      className="btn waves-effect waves-light"
+                      id="login-button"
+                      type="submit"
+                      name="action"
+                      value="submit"
+                    >
+                      Login
+                      <i className="material-icons right">
+                        send
+                      </i>
+                    </button>
                   </div>
                   {this.state.error && <div><h5>{this.state.error}</h5></div>}
                 </form>
@@ -143,18 +164,22 @@ class Login extends React.Component {
                 <Link to="/signup" className=" disabled">New User?</Link>
               </div>
             </div>
-      </li>
-    </ul>
-  </div >
-      
+          </li>
+        </ul>
+      </div >
+
     );
   }
 }
+Login.propTypes = {
+  user: PropTypes.object.isRequired,
+  actions: PropTypes.object.isRequired,
+};
 /**
  *
  *
  * @param {any} state
- * @returns
+ * @returns {null} null
  */
 function mapStateToProps(state) {
   return {
@@ -165,7 +190,7 @@ function mapStateToProps(state) {
  *
  *
  * @param {any} dispatch
- * @returns
+ * @returns {null} null
  */
 function mapDispatchToProps(dispatch) {
   return {
