@@ -1,4 +1,5 @@
 import React from 'react';
+
 const NotAuthorized = () =>
-<h1> You are not authorized to view this page</h1>
+  <h1> You are not authorized to view this page</h1>;
 export default NotAuthorized;
