@@ -1,9 +1,9 @@
 import React from 'react';
 
 const NotFound = () =>
-  <div>
+  (<div>
     <h3>404 page not found</h3>
     <p>We are sorry but the page you are looking for does not exist.</p>
-  </div>;
+  </div>);
 
 export default NotFound;
