@@ -1,9 +1,15 @@
 import React from 'react';
+import PropTypes from 'prop-types';
 import { bindActionCreators } from 'redux';
 import { connect } from 'react-redux';
 import { browserHistory, Link } from 'react-router';
 import * as userActions from '../../actions/userAction';
-
+/**
+ * 
+ * 
+ * @class SignUp
+ * @extends {React.Component}
+ */
 class SignUp extends React.Component {
   /**
    * Creates an instance of SignUp.
@@ -27,31 +33,32 @@ class SignUp extends React.Component {
   /**
    * 
    * 
-   * 
+   * @returns {null} null
    * @memberof SignUp
    */
   componentDidMount() {
-    $(document).ready(function () {
+    $(document).ready(() => {
       $('.slider').slider();
     });
   }
   /**
-   * 
-   * 
-   * @param {any} nextProps 
-   * 
-   * @memberof SignUp
-   */
- componentWillReceiveProps(nextProps) {
-    if(!this.props.user.isAuthenticated && nextProps.user.isAuthenticated) {
-     return browserHistory.push('/home'); 
+ * 
+ * 
+ * @param {any} nextProps 
+ * @returns {state} state  
+ * 
+ * @memberof SignUp
+ */
+  componentWillReceiveProps(nextProps) {
+    if (!this.props.user.isAuthenticated && nextProps.user.isAuthenticated) {
+      return browserHistory.push('/home');
     }
-    return this.setState({error: nextProps.user.error})
+    return this.setState({ error: nextProps.user.error });
   }
   /**
    * 
    * 
-   * 
+   * @returns {null} null  
    * @memberof SignUp
    */
   componentWillUnmount() {
@@ -61,13 +68,13 @@ class SignUp extends React.Component {
    *
    *
    * @param {any} event
-   *
+   * @returns {null} null 
    * @memberof SignUp
    */
   handleSubmit(event) {
     event.preventDefault();
     if (this.passwordConfirmation()) {
-      this.props.actions.signUp(this.state.user)
+      this.props.actions.signUp(this.state.user);
     } else {
       this.setState({ error: 'Passwords do not match' });
     }
@@ -75,7 +82,7 @@ class SignUp extends React.Component {
   /**
    *
    *
-   * @returns
+   * @returns {boolean} boolean
    *
    * @memberof SignUp
    */
@@ -86,7 +93,7 @@ class SignUp extends React.Component {
    *
    *
    * @param {any} event
-   * @returns
+   * @returns {object} state
    *
    * @memberof SignUp
    */
@@ -100,7 +107,7 @@ class SignUp extends React.Component {
   /**
    *
    *
-   * @returns
+   * @returns {object} div
    *
    * @memberof SignUp
    */
@@ -109,65 +116,83 @@ class SignUp extends React.Component {
       <div className="slider fullscreen">
         <ul className="slides">
           <li>
-            <img src="/images/Landing.png" />
+            <img src="/images/Landing.png" alt="" />
             <div className="caption center-align">
-              <h3 id="landing-text" id="app-name">Doc Manager</h3>
+              <h3 id="landing-text">Doc Manager</h3>
               <h3>Sign up</h3>
               <div className="row">
-                <form className="col s12 offset-s3" onSubmit={this.handleSubmit}>
+                <form
+                  className="col s12 offset-s3"
+                  onSubmit={this.handleSubmit}
+                >
                   <div className="row">
                     <div className="input-field col s6">
-                      <input placeholder="Name"
+                      <input
+                        placeholder="Name"
+                        required="required"
                         name="name"
                         id="name"
                         type="text"
                         className="validate"
                         value={this.state.name}
-                        onChange={this.handleChange} />
+                        onChange={this.handleChange}
+                      />
                     </div>
                   </div>
                   <div className="row">
                     <div />
                     <div className="row">
                       <div className="input-field col s3">
-                        <input placeholder="Username"
+                        <input
+                          placeholder="Username"
+                          required="required"
                           name="userName"
                           id="userName"
                           type="text"
                           className="validate"
                           value={this.state.userName}
-                          onChange={this.handleChange} />
+                          onChange={this.handleChange}
+                        />
                       </div>
-                   
+
                       <div className="input-field col s3">
-                        <input placeholder="Email"
+                        <input
+                          placeholder="Email"
+                          required="required"
                           name="email"
                           id="email"
                           type="email"
                           className="validate"
                           value={this.state.email}
-                          onChange={this.handleChange} />
+                          onChange={this.handleChange}
+                        />
                       </div>
                     </div>
                     <div className="row">
                       <div className="input-field col s3">
-                        <input placeholder="Password"
+                        <input
+                          placeholder="Password"
+                          required="required"
                           name="password"
                           id="password"
                           type="password"
                           className="validate"
                           value={this.state.password}
-                          onChange={this.handleChange} />
+                          onChange={this.handleChange}
+                        />
                       </div>
 
                       <div className="input-field col s3">
-                        <input placeholder="Confirm Password"
+                        <input
+                          placeholder="Confirm Password"
+                          required="required"
                           name="confirmPassword"
-                          id="password"
+                          id="confirmPassword"
                           type="password"
                           className="validate"
                           value={this.state.confirmPassword}
-                          onChange={this.handleChange} />
+                          onChange={this.handleChange}
+                        />
                       </div>
                     </div>
                     <div className="row">
@@ -176,17 +201,18 @@ class SignUp extends React.Component {
                       </div>
                     </div>
                     <div className="row">
-                      <button className="btn waves-effect waves-light col s3"
-                        id="button"
+                      <button
+                        className="btn waves-effect waves-light col s3"
+                        id="submit-btn"
                         type="submit"
                         name="action"
-                        value="submit">
+                        value="submit"
+                      >
                         Submit
-                     <i className="material-icons right"> send </i>
+                        <i className="material-icons right"> send </i>
                       </button>
-                      <div>
-                  </div>
-                   <Link to="/" className="col s3">Already A User?</Link>
+                      <div />
+                      <Link to="/" className="col s3">Already A User?</Link>
                     </div>
                     {this.state.error && <div><h5>{this.state.error}</h5></div>}
                   </div>
@@ -199,11 +225,15 @@ class SignUp extends React.Component {
     );
   }
 }
+SignUp.propTypes = {
+  actions: PropTypes.object.isRequired,
+  user: PropTypes.object.isRequired,
+};
 /**
  *
  *
  * @param {any} state
- * @returns
+ * @returns {object} user
  */
 function mapStateToProps(state) {
   return {
@@ -214,7 +244,7 @@ function mapStateToProps(state) {
  *
  *
  * @param {any} dispatch
- * @returns
+ * @returns {object} actions
  */
 function mapDispatchToProps(dispatch) {
   return {
