@@ -1,6 +1,13 @@
 import * as types from '../actions/actionTypes';
 import initialState from './initialState';
-
+/**
+ * 
+ * 
+ * @export
+ * @param {any} [state=initialState.documents] 
+ * @param {any} action 
+ * @returns {object} state
+ */
 export default function docReducer(state = initialState.documents, action) {
   switch (action.type) {
     case types.LOAD_DOCUMENTS_SUCCESS:
@@ -11,7 +18,7 @@ export default function docReducer(state = initialState.documents, action) {
           documents: action.documents.documents.rows,
           metadata: action.documents.metadata
         }
-        );
+      );
     case types.DELETE_DOCUMENTS_SUCCESS:
       return Object.assign(
         {},
@@ -21,6 +28,7 @@ export default function docReducer(state = initialState.documents, action) {
             if (document.id !== action.id) {
               return document;
             }
+            return undefined;
           })
         }
       );
@@ -43,6 +51,8 @@ export default function docReducer(state = initialState.documents, action) {
     case types.CREATE_DOCUMENTS_SUCCESS:
       return Object.assign({}, state, action.document);
 
+    case types.UNAUTHORIZED:
+      return Object.assign({}, state, { error: 'Unauthorized' });
     default:
       return state;
   }
