@@ -2,8 +2,8 @@ export default {
   authUser: {
     isAuthenticated: false
   },
-  users: { users: [], metadata:{} },
-  documents: { documents: [], metadata:{}, document:{} },
-  search: { search: [], metadata:{} }
+  users: { users: [], metadata: {} },
+  documents: { documents: [], metadata: {}, document: {}, error: null },
+  search: { search: [], metadata: {} }
 
 };
