@@ -2,11 +2,17 @@ import reduxImmutableStateInvariant from 'redux-immutable-state-invariant';
 import thunk from 'redux-thunk';
 import { createStore, applyMiddleware } from 'redux';
 import rootReducer from '../reducers';
-
+/**
+ * 
+ * 
+ * @export
+ * @param {any} initialState 
+ * @returns {object} store
+ */
 export default function configureStore(initialState) {
   return createStore(
-      rootReducer,
-      initialState,
-      applyMiddleware(thunk, reduxImmutableStateInvariant())
-    );
+    rootReducer,
+    initialState,
+    applyMiddleware(thunk, reduxImmutableStateInvariant())
+  );
 }
