@@ -1,25 +1,24 @@
 const Documents = require('../models').Documents;
-const Users = require('../models').User;
 
 
 module.exports = {
   /**
    *
    *
-   * @param {any} req
-   * @param {any} res
-   * @returns
+    * @param {any} req request
+   * @param {any} res response
+   * @returns {null} nothing
    */
   create(req, res) {
     const access = req.body.access;
     const title = req.body.title;
-    if(!title || title === '') {
-      return res.status(400).send({ message: 'title required' })
+    if (!title || title === '') {
+      return res.status(400).send({ message: 'title required' });
     } else if (!access) {
-      return res.status(400).send({ message : 'Access required' });
+      return res.status(400).send({ message: 'Access required' });
     }
     if (access !== 'Public' && access !== 'Private' && access !== 'Role') {
-      return res.status(400).send({ message : `No such access right  as ${access} available` });
+      return res.status(400).send({ message: `No such access right  as ${access} available` });
     }
     const userId = req.decoded.id;
     return Documents
@@ -34,9 +33,9 @@ module.exports = {
   /**
    *
    *
-   * @param {any} req
-   * @param {any} res
-   * @returns
+  * @param {any} req request
+   * @param {any} res response
+   * @returns {null} nothing
    */
   findDoc(req, res) {
     const userId = req.decoded.id;
@@ -44,23 +43,26 @@ module.exports = {
       .findById(req.params.id)
       .then((documents) => {
         if (!documents) {
-            return res.status(404).send({
-              message: 'Document Not Found'
+          return res.status(404).send({
+            message: 'Document Not Found'
+          });
+        } else if (documents.userId !== userId
+          && documents.access === 'Private') {
+          return res.status(401)
+            .send({
+              message: 'Not authorized, its not your document'
             });
-        } else if (documents.userId !== userId && documents.access === 'Private') {
-          return res.status(401).send({ message: 'Not authorized, its not your document' });
-        } else {
-          res.status(200)
-            .send({ documents });
         }
+        res.status(200)
+          .send({ documents });
       });
   },
   /**
    *
    *
-   * @param {any} req
-   * @param {any} res
-   * @returns
+   * @param {any} req request
+   * @param {any} res response
+   * @returns {null} nothing
    */
   update(req, res) {
     const id = req.decoded.id;
@@ -68,26 +70,25 @@ module.exports = {
       .findById(req.params.id)
       .then((documents) => {
         if (!documents) {
-            return res.status(404).send({
-              message: 'Document Not Found'
-            });
-          } else if (documents.userId !== id) {
+          return res.status(404).send({
+            message: 'Document Not Found'
+          });
+        } else if (documents.userId !== id) {
           return res.status(401).send({ message: 'Not authorized, its not your document :)' });
-        } else {
-          if (documents.userId === id) {
-            return documents
-              .update(req.body)
-              .then(() => res.status(200).send(documents));
-          }
         }
-      })
+        if (documents.userId === id) {
+          return documents
+            .update(req.body)
+            .then(() => res.status(200).send(documents));
+        }
+      });
   },
   /**
    *
    *
-   * @param {any} req
-   * @param {any} res
-   * @returns
+  * @param {any} req request
+   * @param {any} res response
+   * @returns {null} nothing
    */
   delete(req, res) {
     const userId = req.decoded.id;
@@ -96,19 +97,19 @@ module.exports = {
       .findById(req.params.id)
       .then((documents) => {
         if (!documents) {
-            return res.status(404).send({
-              message: 'Documents Not Found'
-            });
-          } else if (documents.userId !== userId && role !== 1) {
+          return res.status(404).send({
+            message: 'Documents Not Found'
+          });
+        } else if (documents.userId !== userId && role !== 1) {
           res.status(401).send({ message: 'Not authorized, its not your document :)' });
         } else {
           return documents
             .destroy()
             .then(() => res.status(200).send({
               message: 'Documents deleted successfully'
-            }))
+            }));
         }
-      })
+      });
   }
 
 };
