@@ -3,7 +3,7 @@ const jwt = require('jsonwebtoken');
 module.exports = {
   /**
    *
-   *
+   * @returns {null} null
    * @param {any} req
    * @param {any} res
    * @param {any} next
@@ -26,14 +26,13 @@ module.exports = {
    * @param {any} req
    * @param {any} res
    * @param {any} next
-   * @returns
+   * @returns {func} func
    */
   authorizeAdmin(req, res, next) {
     if (req.decoded.role !== 1) {
       return res.status(401).json({ message: 'unathorized' });
-    } else {
-      req.isAdmin = true;
-      return next();
     }
+    req.isAdmin = true;
+    return next();
   }
 };
